Snapshot: Allow snapshots without key-value states
This patch allows snapshots without any key-value states for key-value
blockchain network. It can reduce time and memory for checks by skipping
key-value processing.

Note that snapshots without key-value states will never be used for
further checks that need key-value states.
For this reason, the key-value states are enabled by default when
snapshots are to be saved. Specify the "--skip-key-value" option to
override this behavior.

Signed-off-by: Taku Shimosawa <taku.shimosawa.uf@hitachi.com>